backend/requirements.txt
affine
altgraph
asttokens
attrs
bcrypt
bidict
blinker
Bottleneck 
branca 
Brotli 
cachelib
cairocffi
Cerberus
certifi 
cffi 
charset-normalizer 
click
click-plugins
cligj
cmocean
colorama
config
contourpy 
cssselect2
cycler 
DateTime
debugpy
decorator
defusedxml
executing
Flask
Flask-Caching
flask-cors
Flask-JWT-Extended
Flask-SocketIO
folium 
fonttools 
GDAL 
geopandas 
h11
h2 
hpack 
hyperframe 
idna 
itsdangerous
Jinja2 
joblib 
kiwisolver 
mapclassify 
markdown-it-py
MarkupSafe 
matplotlib
mdurl
munkres
networkx 
numexpr 
numpy 
packaging 
pandas 
pbr
pefile
pillow 
platformdirs
pure_eval
pycparser 
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
pyogrio 
pyparsing 
pyproj 
PySocks 
python-dateutil 
python-dotenv
python-engineio
python-socketio
pytz 
pywin32-ctypes
PyYAML
repoze.lru
requests 
rich
Routes
scikit-learn 
scipy 
setuptools
shapely 
simple-websocket
six 
stack-data
stevedore
threadpoolctl 
tinycss2
tzdata 
unicodedata2 
urllib3 
waitress
wcwidth
webencodings
Werkzeug
wheel
win_inet_pton 
wsproto
XlsxWriter
xyzservices 
zope.interface
zstandard 
